Добавлено обновление меток вне зависимости от нажатой вкладки
