+++ View/NewPinPopup.xaml
@@ -32,7 +32,8 @@
                         Margin="0,0,2,0"
                         WidthRequest="75"
                         HorizontalOptions="Start"
-                        Command="{Binding Source={RelativeSource AncestorType={x:Type vm:SearchViewModel}}, Path=AddImageCommand}"/>
+                        x:DataType="vm:PinManagerViewModel"
+                        Command="{Binding Source={RelativeSource AncestorType={x:Type vm:PinManagerViewModel}}, Path=AddImageCommand}"/>
                 <CollectionView Grid.Column="1"
      ItemsLayout="HorizontalList"
      ItemsSource="{Binding Images}">
