+++ View/PinManagerPage.xaml
@@ -54,6 +54,7 @@
                            EmptyView="Нет доступных туалетов"
                            IsVisible="True"
                            VerticalOptions="Center"
+                           Position="{Binding EditToiletIndex}"
                            x:Name="CarouselCards"
                            ItemsSource="{Binding ActivePinList}">
                 <CarouselView.ItemTemplate>
