+++ View/PinManagerPage.xaml.cs
@@ -3,23 +3,22 @@ using MFASeeker.ViewModel;
 namespace MFASeeker.View;
 public partial class PinManagerPage : ContentPage
 {
-	private PinManagerViewModel _pinManagerVM;
+	//private PinManagerViewModel _pinManagerVM;
     private SwipeView? _previousSwipeView;
     public PinManagerPage(PinManagerViewModel pinManagerVM)
 	{
 		InitializeComponent();
         BindingContext = pinManagerVM;
-        _pinManagerVM = pinManagerVM;
 
 
         //CarouselCards.IsVisible = true;
         //ListCards.IsVisible = false;
     }
-    protected override async void OnAppearing()
-    {
-        base.OnAppearing();
-        await _pinManagerVM.RefreshToiletsCommand.ExecuteAsync(null);
-    }
+    //protected override async void OnAppearing()
+    //{
+    //    base.OnAppearing();
+    //    await _pinManagerVM.RefreshToiletsCommand.ExecuteAsync(null);
+    //}
     private void OnSwipeStarted(object sender, SwipeStartedEventArgs e)
     {
         if (sender as SwipeView is var currentSwipeView)
